refactor(shared): replacement renderer class
