+++ packages/content/src/index.ts
@@ -1,13 +1,8 @@
 import "@worm/shared/vite-env.d.ts";
 
+import { Renderer } from "@worm/shared/src/browser";
 import { runStorageMigrations } from "@worm/shared/src/storage";
 
-import { startContentListeners } from "./lib/listeners";
-import { renderContent } from "./lib/render";
-
-document.addEventListener("readystatechange", () => {
-  renderContent("document state change");
-});
-
 runStorageMigrations();
-startContentListeners();
+
+new Renderer();
