+++ packages/shared/src/browser/index.ts
@@ -3,4 +3,5 @@ export * from "./browser";
 export * from "./connect";
 export * from "./matchers";
 export * from "./popup";
+export * from "./render";
 export * from "./session";
